Refactored ChatbotWebsocket component into two different components: ChatbotWebsocket for ws compatibility and ChatbotHttp for common REST operations.
